move script check from digest article to api crud articles
